people: refactoring
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@5567 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ gpu/people/src/optimized_elec.cpp
@@ -297,7 +297,7 @@ void optimized_elec(const PointCloud<pcl::PointXYZRGB> &cloud, const cv::Mat& sr
 
     int cc = -1;
 
-    double squared_radius = tolerance * tolerance;
+    float squared_radius = tolerance * tolerance;
 
     for(int j = 0; j < sz.height; ++j)
     {
